Splitting user_settings_section in its own component
